# internal/api/openapi.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes = "BearerAuth.Scopes"
)

// Defines values for CreatePlanRequestBillingCycle.
const (
	CreatePlanRequestBillingCycleMonthly CreatePlanRequestBillingCycle = "monthly"
	CreatePlanRequestBillingCycleYearly  CreatePlanRequestBillingCycle = "yearly"
)

// Defines values for CreateSubscriptionRequestBillingCycle.
const (
	CreateSubscriptionRequestBillingCycleMonthly CreateSubscriptionRequestBillingCycle = "monthly"
	CreateSubscriptionRequestBillingCycleYearly  CreateSubscriptionRequestBillingCycle = "yearly"
)

// Defines values for PlanBillingCycle.
const (
	PlanBillingCycleMonthly PlanBillingCycle = "monthly"
	PlanBillingCycleYearly  PlanBillingCycle = "yearly"
)

// Defines values for SubscriptionBillingCycle.
const (
	SubscriptionBillingCycleMonthly SubscriptionBillingCycle = "monthly"
	SubscriptionBillingCycleYearly  SubscriptionBillingCycle = "yearly"
)

// Defines values for UpdatePlanRequestBillingCycle.
const (
	UpdatePlanRequestBillingCycleMonthly UpdatePlanRequestBillingCycle = "monthly"
	UpdatePlanRequestBillingCycleYearly  UpdatePlanRequestBillingCycle = "yearly"
)

// Defines values for UpdateSubscriptionRequestBillingCycle.
const (
	UpdateSubscriptionRequestBillingCycleMonthly UpdateSubscriptionRequestBillingCycle = "monthly"
	UpdateSubscriptionRequestBillingCycleYearly  UpdateSubscriptionRequestBillingCycle = "yearly"
)

// BaseEntity defines model for BaseEntity.
type BaseEntity struct {
	// CreatedAt 作成日時
	CreatedAt time.Time `json:"created_at"`

	// Id エンティティの一意識別子
	Id openapi_types.UUID `json:"id"`

	// UpdatedAt 更新日時
	UpdatedAt time.Time `json:"updated_at"`
}

// CreatePlanRequest defines model for CreatePlanRequest.
type CreatePlanRequest struct {
	// BillingCycle 課金サイクル
	BillingCycle CreatePlanRequestBillingCycle `json:"billing_cycle"`

	// Description 説明（オプション）
	Description *string `json:"description,omitempty"`

	// Features 機能リスト（オプション）
	Features *[]string `json:"features,omitempty"`

	// IsActive 有効フラグ
	IsActive bool `json:"is_active"`

	// Name プラン名
	Name string `json:"name"`

	// Price 価格
	Price float32 `json:"price"`
}

// CreatePlanRequestBillingCycle 課金サイクル
type CreatePlanRequestBillingCycle string

// CreateSubscriptionRequest defines model for CreateSubscriptionRequest.
type CreateSubscriptionRequest struct {
	// BillingCycle 課金サイクル
	BillingCycle CreateSubscriptionRequestBillingCycle `json:"billing_cycle"`

	// Description 説明（オプション）
	Description *string `json:"description,omitempty"`

	// Name サブスクリプション名
	Name string `json:"name"`

	// NextBillingDate 次回課金日
	NextBillingDate openapi_types.Date `json:"next_billing_date"`

	// Price 月額料金
	Price float32 `json:"price"`
}

// CreateSubscriptionRequestBillingCycle 課金サイクル
type CreateSubscriptionRequestBillingCycle string

// Error defines model for Error.
type Error struct {
	Code    *string `json:"code,omitempty"`
	Message *string `json:"message,omitempty"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Email    openapi_types.Email `json:"email"`
	Password string              `json:"password"`
}

// LoginResponse defines model for LoginResponse.
type LoginResponse struct {
	// Token JWTトークン
	Token *string `json:"token,omitempty"`
}

// Plan defines model for Plan.
type Plan struct {
	// BillingCycle 課金サイクル
	BillingCycle PlanBillingCycle `json:"billing_cycle"`

	// CreatedAt 作成日時
	CreatedAt time.Time `json:"created_at"`

	// Description 説明（オプション）
	Description *string `json:"description,omitempty"`

	// Features 機能リスト（オプション）
	Features *[]string `json:"features,omitempty"`

	// Id エンティティの一意識別子
	Id openapi_types.UUID `json:"id"`

	// IsActive 有効フラグ
	IsActive bool `json:"is_active"`

	// Name プラン名
	Name string `json:"name"`

	// Price 価格
	Price float32 `json:"price"`

	// UpdatedAt 更新日時
	UpdatedAt time.Time `json:"updated_at"`
}

// PlanBillingCycle 課金サイクル
type PlanBillingCycle string

// SignUpRequest defines model for SignUpRequest.
type SignUpRequest struct {
	// Email メールアドレス
	Email openapi_types.Email `json:"email"`

	// Name ユーザー名
	Name string `json:"name"`

	// Password パスワード（8文字以上）
	Password string `json:"password"`
}

// SignUpResponse defines model for SignUpResponse.
type SignUpResponse struct {
	// Token JWTトークン
	Token  *string             `json:"token,omitempty"`
	UserId *openapi_types.UUID `json:"user_id,omitempty"`
}

// Subscription defines model for Subscription.
type Subscription struct {
	// BillingCycle 課金サイクル
	BillingCycle SubscriptionBillingCycle `json:"billing_cycle"`

	// CreatedAt 作成日時
	CreatedAt time.Time `json:"created_at"`

	// Description 説明（オプション）
	Description *string `json:"description,omitempty"`

	// Id エンティティの一意識別子
	Id openapi_types.UUID `json:"id"`

	// Name サブスクリプション名
	Name string `json:"name"`

	// NextBillingDate 次回課金日
	NextBillingDate openapi_types.Date `json:"next_billing_date"`

	// Price 月額料金
	Price float32 `json:"price"`

	// UpdatedAt 更新日時
	UpdatedAt time.Time `json:"updated_at"`

	// UserId ユーザーID
	UserId openapi_types.UUID `json:"user_id"`
}

// SubscriptionBillingCycle 課金サイクル
type SubscriptionBillingCycle string

// UpdatePlanRequest defines model for UpdatePlanRequest.
type UpdatePlanRequest struct {
	// BillingCycle 課金サイクル
	BillingCycle UpdatePlanRequestBillingCycle `json:"billing_cycle"`

	// Description 説明（オプション）
	Description *string `json:"description,omitempty"`

	// Features 機能リスト（オプション）
	Features *[]string `json:"features,omitempty"`

	// IsActive 有効フラグ
	IsActive bool `json:"is_active"`

	// Name プラン名
	Name string `json:"name"`

	// Price 価格
	Price float32 `json:"price"`
}

// UpdatePlanRequestBillingCycle 課金サイクル
type UpdatePlanRequestBillingCycle string

// UpdateSubscriptionRequest defines model for UpdateSubscriptionRequest.
type UpdateSubscriptionRequest struct {
	// BillingCycle 課金サイクル
	BillingCycle UpdateSubscriptionRequestBillingCycle `json:"billing_cycle"`

	// Description 説明（オプション）
	Description *string `json:"description,omitempty"`

	// Name サブスクリプション名
	Name string `json:"name"`

	// NextBillingDate 次回課金日
	NextBillingDate openapi_types.Date `json:"next_billing_date"`

	// Price 月額料金
	Price float32 `json:"price"`
}

// UpdateSubscriptionRequestBillingCycle 課金サイクル
type UpdateSubscriptionRequestBillingCycle string

// LoginJSONRequestBody defines body for Login for application/json ContentType.
type LoginJSONRequestBody = LoginRequest

// SignUpJSONRequestBody defines body for SignUp for application/json ContentType.
type SignUpJSONRequestBody = SignUpRequest

// CreatePlanJSONRequestBody defines body for CreatePlan for application/json ContentType.
type CreatePlanJSONRequestBody = CreatePlanRequest

// UpdatePlanJSONRequestBody defines body for UpdatePlan for application/json ContentType.
type UpdatePlanJSONRequestBody = UpdatePlanRequest

// CreateSubscriptionJSONRequestBody defines body for CreateSubscription for application/json ContentType.
type CreateSubscriptionJSONRequestBody = CreateSubscriptionRequest

// UpdateSubscriptionJSONRequestBody defines body for UpdateSubscription for application/json ContentType.
type UpdateSubscriptionJSONRequestBody = UpdateSubscriptionRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// ログイン
	// (POST /auth/login)
	Login(ctx echo.Context) error
	// ユーザー登録
	// (POST /auth/signup)
	SignUp(ctx echo.Context) error
	// プラン一覧取得
	// (GET /plans)
	ListPlans(ctx echo.Context) error
	// プラン作成
	// (POST /plans)
	CreatePlan(ctx echo.Context) error
	// プラン削除
	// (DELETE /plans/{id})
	DeletePlan(ctx echo.Context, id openapi_types.UUID) error
	// プラン詳細取得
	// (GET /plans/{id})
	GetPlan(ctx echo.Context, id openapi_types.UUID) error
	// プラン更新
	// (PUT /plans/{id})
	UpdatePlan(ctx echo.Context, id openapi_types.UUID) error
	// サブスクリプション一覧取得
	// (GET /subscriptions)
	ListSubscriptions(ctx echo.Context) error
	// サブスクリプション作成
	// (POST /subscriptions)
	CreateSubscription(ctx echo.Context) error
	// サブスクリプション詳細取得
	// (GET /subscriptions/{id})
	GetSubscription(ctx echo.Context, id openapi_types.UUID) error
	// サブスクリプション更新
	// (PUT /subscriptions/{id})
	UpdateSubscription(ctx echo.Context, id openapi_types.UUID) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// Login converts echo context to params.
func (w *ServerInterfaceWrapper) Login(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.Login(ctx)
	return err
}

// SignUp converts echo context to params.
func (w *ServerInterfaceWrapper) SignUp(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.SignUp(ctx)
	return err
}

// ListPlans converts echo context to params.
func (w *ServerInterfaceWrapper) ListPlans(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ListPlans(ctx)
	return err
}

// CreatePlan converts echo context to params.
func (w *ServerInterfaceWrapper) CreatePlan(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreatePlan(ctx)
	return err
}

// DeletePlan converts echo context to params.
func (w *ServerInterfaceWrapper) DeletePlan(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeletePlan(ctx, id)
	return err
}

// GetPlan converts echo context to params.
func (w *ServerInterfaceWrapper) GetPlan(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetPlan(ctx, id)
	return err
}

// UpdatePlan converts echo context to params.
func (w *ServerInterfaceWrapper) UpdatePlan(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.UpdatePlan(ctx, id)
	return err
}

// ListSubscriptions converts echo context to params.
func (w *ServerInterfaceWrapper) ListSubscriptions(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ListSubscriptions(ctx)
	return err
}

// CreateSubscription converts echo context to params.
func (w *ServerInterfaceWrapper) CreateSubscription(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateSubscription(ctx)
	return err
}

// GetSubscription converts echo context to params.
func (w *ServerInterfaceWrapper) GetSubscription(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetSubscription(ctx, id)
	return err
}

// UpdateSubscription converts echo context to params.
func (w *ServerInterfaceWrapper) UpdateSubscription(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.UpdateSubscription(ctx, id)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/auth/login", wrapper.Login)
	router.POST(baseURL+"/auth/signup", wrapper.SignUp)
	router.GET(baseURL+"/plans", wrapper.ListPlans)
	router.POST(baseURL+"/plans", wrapper.CreatePlan)
	router.DELETE(baseURL+"/plans/:id", wrapper.DeletePlan)
	router.GET(baseURL+"/plans/:id", wrapper.GetPlan)
	router.PUT(baseURL+"/plans/:id", wrapper.UpdatePlan)
	router.GET(baseURL+"/subscriptions", wrapper.ListSubscriptions)
	router.POST(baseURL+"/subscriptions", wrapper.CreateSubscription)
	router.GET(baseURL+"/subscriptions/:id", wrapper.GetSubscription)
	router.PUT(baseURL+"/subscriptions/:id", wrapper.UpdateSubscription)

}